requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==3.0.1
Django==4.2.20
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.5.1
gunicorn==23.0.0
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.5.0
whitenoise==6.9.0
